+++ components/camel-servicenow/camel-servicenow-component/src/test/java/org/apache/camel/component/servicenow/ServiceNowITSupport.java
@@ -26,16 +26,16 @@ import org.apache.camel.util.ObjectHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-class ServiceNowTestSupport extends CamelTestSupport {
-    protected static final Logger LOGGER = LoggerFactory.getLogger(ServiceNowTestSupport.class);
+class ServiceNowITSupport extends CamelTestSupport {
+    protected static final Logger LOGGER = LoggerFactory.getLogger(ServiceNowITSupport.class);
 
     private final boolean setUpComponent;
 
-    ServiceNowTestSupport() {
+    ServiceNowITSupport() {
         this(true);
     }
 
-    ServiceNowTestSupport(boolean setUpComponent) {
+    ServiceNowITSupport(boolean setUpComponent) {
         this.setUpComponent = setUpComponent;
     }
