+++ components/camel-servicenow/camel-servicenow-component/src/test/resources/log4j2.properties
@@ -29,7 +29,7 @@ appender.out.layout.pattern = %d{HH:mm:ss.SSS} [%-15.15t] %-30.30c{1} %-5p %m%n
 logger.servicenow.name = org.apache.camel.component.servicenow
 logger.servicenow.level = DEBUG
 
-logger.metadata.name = org.apache.camel.component.servicenow.ServiceNowMetaDataExtensionTest
+logger.metadata.name = org.apache.camel.component.servicenow.ServiceNowMetaDataExtensionIT
 logger.metadata.level = DEBUG
 #logger.metadata.additivity = true
 #logger.metadata.appenderRef.metadata.ref = out
